Added new command
